refactor(core): extract WorkerIO to framework layer (sync API preserved)
Spec section 5 (docs/specs/concurrency-capability-framework.md) calls for
a generic worker_io abstraction reusable beyond TRT-Edge-LLM TTS. This
commit is the pure-extraction step (P1a):

  * New module app/core/worker_io.py with the WorkerIO class +
    WorkerExitError, copied verbatim from app/backends/jetson/
    trt_edge_llm_tts.py:482..628 with names un-privated
    (_WorkerIO -> WorkerIO). The private name is preserved as a
    module-level alias for any straggler reference.
  * trt_edge_llm_tts.py imports WorkerIO/WorkerExitError from
    app.core.worker_io; the local class bodies (147 lines) are
    deleted. Type annotation + constructor call updated to the
    public name.
  * app/main.py /health import (888) switched from the now-removed
    private TTS-module symbol to the framework module.

API surface is unchanged: sync request() generator + GeneratorExit
cancel semantics from commit 64185fa preserved. The async migration
to send_request() is the next commit (P1b). Zero behavior change.

Existing TTS test suite delta: 11 pre-existing failures on Mac at
07cbcf2 -> 10 here, no new failures introduced.
